# apps/web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backtest Research Web API",
    "description": "Read-only P0 backend-for-frontend. Stored catalog summaries are returned without metric or decision recomputation.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/runs": {
      "get": {
        "summary": "List Runs",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy Id"
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timeframe"
            }
          },
          {
            "name": "exchange",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exchange"
            }
          },
          {
            "name": "market_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market Type"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "decision_route",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision Route"
            }
          },
          {
            "name": "gate_passed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gate Passed"
            }
          },
          {
            "name": "sweep_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sweep Id"
            }
          },
          {
            "name": "config_hash",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Config Hash"
            }
          },
          {
            "name": "created_at_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created At From"
            }
          },
          {
            "name": "created_at_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created At To"
            }
          },
          {
            "name": "period_start_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period Start From"
            }
          },
          {
            "name": "period_start_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period Start To"
            }
          },
          {
            "name": "period_end_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period End From"
            }
          },
          {
            "name": "period_end_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period End To"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "-created_at",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunHeader"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/summary": {
      "get": {
        "summary": "Get Run Summary",
        "operationId": "get_run_summary_api_v1_runs__run_id__summary_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummaryResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "JsonValue": {},
      "Page": {
        "properties": {
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "limit",
          "offset",
          "total",
          "has_more"
        ],
        "title": "Page",
        "description": "Offset page metadata."
      },
      "RunHeader": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "run_seq": {
            "type": "integer",
            "title": "Run Seq"
          },
          "run_name": {
            "type": "string",
            "title": "Run Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name"
          },
          "strategy_version": {
            "type": "string",
            "title": "Strategy Version"
          },
          "params_json": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "params_schema_version": {
            "type": "string",
            "title": "Params Schema Version"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "market_type": {
            "type": "string",
            "title": "Market Type"
          },
          "period_start": {
            "type": "string",
            "format": "date-time",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date-time",
            "title": "Period End"
          },
          "warmup_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warmup Start"
          },
          "warmup_candles": {
            "type": "integer",
            "title": "Warmup Candles"
          },
          "data_source": {
            "type": "string",
            "title": "Data Source"
          },
          "indicator_mode": {
            "type": "string",
            "title": "Indicator Mode"
          },
          "trigger_feed": {
            "type": "string",
            "title": "Trigger Feed"
          },
          "fill_timing": {
            "type": "string",
            "title": "Fill Timing"
          },
          "initial_capital": {
            "type": "string",
            "format": "decimal",
            "title": "Initial Capital",
            "description": "Exact PostgreSQL NUMERIC value serialized without floating-point conversion."
          },
          "sizing_method": {
            "type": "string",
            "title": "Sizing Method"
          },
          "risk_per_trade": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal",
                "description": "Exact PostgreSQL NUMERIC value serialized without floating-point conversion."
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Per Trade"
          },
          "position_size_pct": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal",
                "description": "Exact PostgreSQL NUMERIC value serialized without floating-point conversion."
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Size Pct"
          },
          "framework_compliant": {
            "type": "boolean",
            "title": "Framework Compliant"
          },
          "cost_values_json": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "engine_version": {
            "type": "string",
            "title": "Engine Version"
          },
          "core_lib_version": {
            "type": "string",
            "title": "Core Lib Version"
          },
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          },
          "profile_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Ref"
          },
          "strategy_profile_json": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "envelope_status_declared": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Envelope Status Declared"
          },
          "sweep_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sweep Id"
          },
          "fold_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fold Label"
          },
          "evidence_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Path"
          },
          "evidence_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Hash"
          },
          "evidence_retained": {
            "type": "boolean",
            "title": "Evidence Retained"
          },
          "evidence_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Expires At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "resolved_indicators_json": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "source_data_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Data Hash"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "run_seq",
          "run_name",
          "status",
          "strategy_id",
          "strategy_name",
          "strategy_version",
          "params_json",
          "params_schema_version",
          "symbol",
          "exchange",
          "timeframe",
          "market_type",
          "period_start",
          "period_end",
          "warmup_start",
          "warmup_candles",
          "data_source",
          "indicator_mode",
          "trigger_feed",
          "fill_timing",
          "initial_capital",
          "sizing_method",
          "risk_per_trade",
          "position_size_pct",
          "framework_compliant",
          "cost_values_json",
          "seed",
          "engine_version",
          "core_lib_version",
          "config_hash",
          "profile_ref",
          "strategy_profile_json",
          "envelope_status_declared",
          "sweep_id",
          "fold_label",
          "evidence_path",
          "evidence_hash",
          "evidence_retained",
          "evidence_expires_at",
          "error_message",
          "started_at",
          "finished_at",
          "created_at",
          "resolved_indicators_json",
          "source_data_hash"
        ],
        "title": "RunHeader",
        "description": "Reproducibility header mirroring every backtest_run column."
      },
      "RunListItem": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "run_name": {
            "type": "string",
            "title": "Run Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "market_type": {
            "type": "string",
            "title": "Market Type"
          },
          "period_start": {
            "type": "string",
            "format": "date-time",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date-time",
            "title": "Period End"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "sweep_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sweep Id"
          },
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          },
          "trade_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Count"
          },
          "pf": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pf"
          },
          "sortino": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sortino"
          },
          "calmar_or_mar": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calmar Or Mar"
          },
          "sqn": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sqn"
          },
          "mdd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mdd"
          },
          "ror": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ror"
          },
          "win_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Win Rate"
          },
          "net_pnl_total": {
            "anyOf": [
              {
                "type": "string",
                "format": "decimal",
                "description": "Exact PostgreSQL NUMERIC value serialized without floating-point conversion."
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Pnl Total"
          },
          "gate_verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Verdict"
          },
          "decision_route": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Route"
          },
          "integrity_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Integrity Status"
          },
          "data_coverage_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Coverage Ratio"
          },
          "summary_present": {
            "type": "boolean",
            "title": "Summary Present"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "run_name",
          "status",
          "strategy_id",
          "strategy_name",
          "symbol",
          "exchange",
          "timeframe",
          "market_type",
          "period_start",
          "period_end",
          "created_at",
          "sweep_id",
          "config_hash",
          "trade_count",
          "pf",
          "sortino",
          "calmar_or_mar",
          "sqn",
          "mdd",
          "ror",
          "win_rate",
          "net_pnl_total",
          "gate_verdict",
          "decision_route",
          "integrity_status",
          "data_coverage_ratio",
          "summary_present"
        ],
        "title": "RunListItem",
        "description": "Thin catalog row with an optional stored summary."
      },
      "RunListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RunListItem"
            },
            "type": "array",
            "title": "Data"
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          }
        },
        "type": "object",
        "required": [
          "data",
          "page"
        ],
        "title": "RunListResponse",
        "description": "Catalog collection envelope."
      },
...
      "RunSummaryResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "run_status": {
            "type": "string",
            "title": "Run Status"
          },
          "summary_status": {
            "type": "string",
            "enum": [
              "available",
              "pending",
              "failed",
              "orphaned",
              "missing"
            ],
            "title": "Summary Status"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "run_status",
          "summary_status",
          "summary"
        ],
        "title": "RunSummaryResponse",
        "description": "0..1 stored summary with a status-derived absence explanation."
      },
...
    }
  }
}